Update model.py
Tentative edit that makes the code a lot faster but interpolating over the filter, and then just using the filter function values at the wavelengths. Previous runtime was about 8 minutes. Now it's about 2. It now seems to be limited by the file-loading time (I speed tested it again to see).