Create volfun.py
Borrowed and edited Leonidas' code.
The output is a three column array where the first tells you the magnification, the second is the total volume for magnifications greater than the number in the first column and the third is the second minus the areas of the sky covered by bright sources (using the seg maps). The seg maps have not quite been fully incorporated yet. Note that to actually get the volume at each magnification, it's necessary to take the difference. 
You can feed a redshift, minmu and maxmu to the function.